rewrite the entire bot for slash commands
