+++ .gitignore
@@ -8,3 +8,4 @@ CHOpt.exe
 prefixes.json
 json
 midi_files
+**/__pycache__
